+++ src/ray.rs
@@ -10,7 +10,7 @@ impl Ray {
     pub fn new(origin: Vec3, direction: Vec3) -> Ray {
         Ray { origin, direction }
     }
-    pub fn at(self, t: f64) -> Vec3 {
+    pub fn at(self, t: f32) -> Vec3 {
         self.origin + (self.direction * t)
     }
 }
